StoryService: Use StoryRequest to update stories
Story type is no longer being used for updating stories. Now it is
StoryRequest, which contains pointers for correct encoding of empty
sliced. When getting a story from PT API, though, a Story is returned,
which does not contain pointers unless necessary.

StoryRequest also only contains fields that can be updated, or where it
actually makes sense to update them.