StoryService: Do not use Iterate() for List()
Seems like iterating over stories is broken when a filter is supplied. Pivotal
Tracker API is not always returning the same results for the same filter and
pagination setup. Looks like the underlying list of stories generated by the API
is not sorted, then the same pagination setup is associated with a different set
of stories. Will see what PT guys reply, but right now we need to get this
working correctly.

So, List() is not using cursor.all(), which sends an HTTP request to get the
total number of items, then sends another HTTP request to retrive all the items
at once.